Introduce JvmTaskExecutor for pluggable compilation strategies
- Introduce JvmTaskExecutor abstraction to allow legacy (KotlinJvmTaskExecutor) and BTA-based implementations (comes later)
- Extract JvmCompilationTask-manipulation helpers that will be reused by the BTA JvmTaskExecutor implementation in a separate TaskArtifacts.kt file
- CompilationTask.kt keeps helpers needed for the legacy KotlinJvmTaskExecutor implementation
