+++ src/main/kotlin/io/bazel/kotlin/builder/tasks/JvmTaskExecutor.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2026 The Bazel Authors. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
+package io.bazel.kotlin.builder.tasks
+
+import io.bazel.kotlin.builder.toolchain.CompilationTaskContext
+import io.bazel.kotlin.model.JvmCompilationTask
+
+/**
+ * Executes one JVM compilation task inside the persistent worker.
+ *
+ * [KotlinBuilder] parses each work request into a [JvmCompilationTask] and executes it through
+ * this interface; implementations define the complete compilation strategy for the task.
+ */
+interface JvmTaskExecutor {
+  fun execute(
+    context: CompilationTaskContext,
+    task: JvmCompilationTask,
+  )
+}
