+++ src/main/kotlin/io/bazel/kotlin/builder/tasks/jvm/TaskArtifacts.kt
@@ -1,0 +1,247 @@
+/*
+ * Copyright 2026 The Bazel Authors. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
+
+// Task artifacts-related helpers for the JvmCompilationTask protocol buffer
+package io.bazel.kotlin.builder.tasks.jvm
+
+import io.bazel.kotlin.builder.toolchain.CompilationTaskContext
+import io.bazel.kotlin.builder.utils.IS_JVM_SOURCE_FILE
+import io.bazel.kotlin.builder.utils.bazelRuleKind
+import io.bazel.kotlin.builder.utils.jars.JarCreator
+import io.bazel.kotlin.builder.utils.jars.JarHelper.Companion.MANIFEST_DIR
+import io.bazel.kotlin.builder.utils.jars.SourceJarExtractor
+import io.bazel.kotlin.builder.utils.partitionJvmSources
+import io.bazel.kotlin.model.JvmCompilationTask
+import io.bazel.kotlin.model.JvmCompilationTask.Directories
+import java.nio.file.Files
+import java.nio.file.Files.isDirectory
+import java.nio.file.Files.walk
+import java.nio.file.Paths
+import java.nio.file.StandardCopyOption
+import java.util.stream.Stream
+import kotlin.io.path.exists
+
+internal const val SOURCE_JARS_DIR = "_srcjars"
+
+internal fun JvmCompilationTask.preProcessingSteps(
+  context: CompilationTaskContext,
+): JvmCompilationTask = context.execute("expand sources") { expandWithSourceJarSources() }
+
+/**
+ * Produce the primary output jar.
+ */
+internal fun JvmCompilationTask.createOutputJar() =
+  JarCreator(
+    path = Paths.get(outputs.jar),
+    normalize = true,
+    verbose = false,
+  ).also {
+    it.addDirectory(Paths.get(directories.classes))
+    it.addDirectory(Paths.get(directories.javaClasses))
+    it.addDirectory(Paths.get(directories.generatedClasses))
+    it.setJarOwner(info.label, info.bazelRuleKind)
+    it.execute()
+  }
+
+/**
+ * Produce the ABI output jar.
+ */
+internal fun JvmCompilationTask.createAbiJar() =
+  JarCreator(
+    path = Paths.get(outputs.abijar),
+    normalize = true,
+    verbose = false,
+  ).also {
+    it.addDirectory(Paths.get(directories.abiClasses))
+    it.addDirectory(Paths.get(directories.generatedClasses))
+    it.setJarOwner(info.label, info.bazelRuleKind)
+    it.execute()
+  }
+
+/**
+ * Produce a jar of sources generated by KAPT.
+ */
+internal fun JvmCompilationTask.createGeneratedJavaSrcJar() {
+  JarCreator(
+    path = Paths.get(outputs.generatedJavaSrcJar),
+    normalize = true,
+    verbose = false,
+  ).also {
+    it.addDirectory(Paths.get(directories.generatedJavaSources))
+    it.setJarOwner(info.label, info.bazelRuleKind)
+    it.execute()
+  }
+}
+
+/**
+ * Produce a stub jar of classes generated by KAPT.
+ */
+internal fun JvmCompilationTask.createGeneratedStubJar() {
+  JarCreator(
+    path = Paths.get(outputs.generatedJavaStubJar),
+    normalize = true,
+    verbose = false,
+  ).also {
+    it.addDirectory(Paths.get(directories.incrementalData))
+    it.setJarOwner(info.label, info.bazelRuleKind)
+    it.execute()
+  }
+}
+
+/**
+ * Produce a jar of classes generated by KAPT.
+ */
+internal fun JvmCompilationTask.createGeneratedClassJar() {
+  JarCreator(
+    path = Paths.get(outputs.generatedClassJar),
+    normalize = true,
+    verbose = false,
+  ).also {
+    it.addDirectory(Paths.get(directories.generatedClasses))
+    it.setJarOwner(info.label, info.bazelRuleKind)
+    it.execute()
+  }
+}
+
+internal fun JvmCompilationTask.createGeneratedKspKotlinSrcJar() {
+  JarCreator(
+    path = Paths.get(outputs.generatedKspSrcJar),
+    normalize = true,
+    verbose = false,
+  ).also {
+    it.addDirectory(Paths.get(directories.generatedSources))
+    it.addDirectory(Paths.get(directories.generatedJavaSources))
+    it.setJarOwner(info.label, info.bazelRuleKind)
+    it.execute()
+  }
+}
+
+/**
+ * Produce a jar of classes generated by KSP.
+ */
+internal fun JvmCompilationTask.createdGeneratedKspClassesJar() {
+  JarCreator(
+    path = Paths.get(outputs.generatedKspClassesJar),
+    normalize = true,
+    verbose = false,
+  ).also {
+    it.addDirectory(Paths.get(directories.generatedClasses))
+    it.setJarOwner(info.label, info.bazelRuleKind)
+    it.execute()
+  }
+}
+
+/**
+ * If any srcjars were provided expand the jars sources and create a new [JvmCompilationTask] with the
+ * Java, Kotlin sources and META folder merged in.
+ */
+internal fun JvmCompilationTask.expandWithSourceJarSources(): JvmCompilationTask =
+  if (inputs.sourceJarsList.isEmpty()) {
+    this
+  } else {
+    expandWithSources(
+      SourceJarExtractor(
+        destDir = Paths.get(directories.temp).resolve(SOURCE_JARS_DIR),
+        fileMatcher = { str: String -> IS_JVM_SOURCE_FILE.test(str) || "/$MANIFEST_DIR" in str },
+      ).also {
+        it.jarFiles.addAll(inputs.sourceJarsList.map { p -> Paths.get(p) })
+        it.execute()
+      }.sourcesList
+        .iterator(),
+    )
+  }
+
+internal val Directories.stubs
+  get() =
+    Files
+      .createDirectories(
+        Paths
+          .get(temp)
+          .resolve("stubs"),
+      ).toString()
+internal val Directories.incrementalData
+  get() =
+    Files
+      .createDirectories(
+        Paths
+          .get(temp)
+          .resolve("incrementalData"),
+      ).toString()
+
+/**
+ * Create a new [JvmCompilationTask] with sources found in the generatedSources directory. This should be run after
+ * annotation processors have been run.
+ */
+fun JvmCompilationTask.expandWithGeneratedSources(): JvmCompilationTask =
+  expandWithSources(
+    Stream
+      .of(directories.generatedSources, directories.generatedJavaSources)
+      .map { s -> Paths.get(s) }
+      .flatMap { p -> walk(p) }
+      .filter { !isDirectory(it) }
+      .map { it.toString() }
+      .distinct()
+      .iterator(),
+  )
+
+private fun JvmCompilationTask.expandWithSources(sources: Iterator<String>): JvmCompilationTask =
+  updateBuilder { builder ->
+    sources
+      .copyManifestFilesToGeneratedClasses(directories)
+      .filterOutNonCompilableSources()
+      .partitionJvmSources(
+        { builder.inputsBuilder.addKotlinSources(it) },
+        { builder.inputsBuilder.addJavaSources(it) },
+      )
+  }
+
+private fun JvmCompilationTask.updateBuilder(
+  block: (JvmCompilationTask.Builder) -> Unit,
+): JvmCompilationTask =
+  toBuilder().let {
+    block(it)
+    it.build()
+  }
+
+/**
+ * Copy generated manifest files from KSP task into generated folder
+ */
+internal fun Iterator<String>.copyManifestFilesToGeneratedClasses(
+  directories: Directories,
+): Iterator<String> {
+  val result = mutableSetOf<String>()
+  this.forEach {
+    if ("/$MANIFEST_DIR" in it) {
+      val path = Paths.get(it)
+      val srcJarsPath = Paths.get(directories.temp, SOURCE_JARS_DIR)
+      if (srcJarsPath.exists()) {
+        val relativePath = srcJarsPath.relativize(path)
+        val destPath = Paths.get(directories.generatedClasses).resolve(relativePath)
+        destPath.parent.toFile().mkdirs()
+        Files.copy(path, destPath, StandardCopyOption.REPLACE_EXISTING)
+      }
+    }
+    result.add(it)
+  }
+  return result.iterator()
+}
+
+/**
+ * Only keep java and kotlin files for the iterator. Filter out all other non-compilable files.
+ */
+private fun Iterator<String>.filterOutNonCompilableSources(): Iterator<String> =
+  asSequence().filter { it.endsWith(".kt") || it.endsWith(".java") }.iterator()
